third_party/blink/web_tests/fast/events/touch/touch-fractional-coordinates-expected.txt

Tests non-integer TouchEvent co-ordinates and radii

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

  
PASS successfullyParsed is true

TEST COMPLETE

Testing simple fractional touch
PASS eventCount is 1
PASS lastEvent.type is "touchstart"
PASS lastEvent.changedTouches[0].screenX is within 0.00001 of 30.33
PASS lastEvent.changedTouches[0].screenY is within 0.00001 of 4.5
PASS lastEvent.changedTouches[0].clientX is within 0.00001 of 30.33
PASS lastEvent.changedTouches[0].clientY is within 0.00001 of 4.5
PASS lastEvent.changedTouches[0].pageX is within 0.00001 of 33.33
PASS lastEvent.changedTouches[0].pageY is within 0.00001 of 14.5
PASS lastEvent.changedTouches[0].radiusX is within 0.00001 of 5.2
PASS lastEvent.changedTouches[0].radiusY is within 0.00001 of 6.3


Testing fractional touch inside simple iframe
PASS eventCount is 1
PASS lastEvent.type is "touchstart"
PASS lastEvent.changedTouches[0].screenX is within 0.00001 of 11.5
PASS lastEvent.changedTouches[0].screenY is within 0.00001 of 92.2
PASS lastEvent.changedTouches[0].clientX is within 0.00001 of 4.5
PASS lastEvent.changedTouches[0].clientY is within 0.00001 of 2.2
PASS lastEvent.changedTouches[0].pageX is within 0.00001 of 4.5
PASS lastEvent.changedTouches[0].pageY is within 0.00001 of 2.2


Testing fractional touch inside rotated iframe
PASS eventCount is 1
PASS lastEvent.type is "touchstart"
PASS lastEvent.changedTouches[0].screenX is within 0.00001 of 115.5
PASS lastEvent.changedTouches[0].screenY is within 0.00001 of 92.2
PASS lastEvent.changedTouches[0].clientX is within 0.00001 of 95.5
PASS lastEvent.changedTouches[0].clientY is within 0.00001 of 97.8
PASS lastEvent.changedTouches[0].pageX is within 0.00001 of 95.5
PASS lastEvent.changedTouches[0].pageY is within 0.00001 of 97.8


Testing fractional touch inside scaled iframe
PASS eventCount is 1
PASS lastEvent.type is "touchstart"
PASS lastEvent.changedTouches[0].screenX is within 0.00001 of 244.5
PASS lastEvent.changedTouches[0].screenY is within 0.00001 of 97.2
PASS lastEvent.changedTouches[0].clientX is within 0.00001 of 2.25
PASS lastEvent.changedTouches[0].clientY is within 0.00001 of 1.1
PASS lastEvent.changedTouches[0].pageX is within 0.00001 of 2.25
PASS lastEvent.changedTouches[0].pageY is within 0.00001 of 1.1
